+++ src/problem/constr.c
@@ -75,6 +75,44 @@ struct Constr {
   Constr* next; /**< @brief List of constraints */
 };
 
+void CONSTR_finalize_structure_of_Hessians(Constr* c) {
+
+  // Local variables
+  int* Hi_comb;
+  int* Hj_comb;
+  int H_nnz_comb;
+  Mat* H_array;
+  int* Hi;
+  int* Hj;
+  int temp;
+  int k;
+  int m;
+  
+  // Check
+  if (!c)
+    return;
+
+  // Ensure lower triangular and save struct of H comb
+  H_nnz_comb = 0;
+  H_array = CONSTR_get_H_array(c);
+  Hi_comb = MAT_get_row_array(CONSTR_get_H_combined(c));
+  Hj_comb = MAT_get_col_array(CONSTR_get_H_combined(c));
+  for (k = 0; k < CONSTR_get_H_array_size(c); k++) {
+    Hi = MAT_get_row_array(MAT_array_get(H_array,k));
+    Hj = MAT_get_col_array(MAT_array_get(H_array,k));
+    for (m = 0; m < MAT_get_nnz(MAT_array_get(H_array,k)); m++) {
+      if (Hi[m] < Hj[m]) {
+	temp = Hi[m];
+	Hi[m] = Hj[m];
+	Hj[m] = temp;
+      }
+      Hi_comb[H_nnz_comb] = Hi[m];
+      Hj_comb[H_nnz_comb] = Hj[m];
+      H_nnz_comb++;
+    }
+  }
+}
+
 int CONSTR_get_num_extra_vars(Constr* c) {
   if (c)
     return c->num_extra_vars;
@@ -411,6 +449,12 @@ Constr* CONSTR_get_next(Constr* c) {
     return c->next;
   else
     return NULL;
+}
+
+void CONSTR_list_finalize_structure_of_Hessians(Constr* clist) {
+  Constr* cc;
+  for (cc = clist; cc != NULL; cc = CONSTR_get_next(cc))
+    CONSTR_finalize_structure_of_Hessians(cc);
 }
 
 void CONSTR_list_clear_error(Constr* clist) {
@@ -793,6 +837,7 @@ void CONSTR_analyze(Constr* c) {
     for (i = 0; i < NET_get_num_branches(net); i++)
       CONSTR_analyze_step(c,NET_get_branch(net,i),t);
   }
+  CONSTR_finalize_structure_of_Hessians(c);
 }
 
 void CONSTR_analyze_step(Constr* c, Branch* br, int t) {