+++ src/problem/constr/constr_ACPF.c
@@ -495,14 +495,9 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
   int* J_nnz;
   int* H_nnz;
   int H_nnz_val;
-  int H_nnz_comb;
   char* bus_counted;
   Mat* H_array;
   Mat* H[2];
-  int* Hi;
-  int* Hj;
-  int* Hi_comb;
-  int* Hj_comb;
   int bus_index_t[2];
   int w_index[2];
   int v_index[2];
@@ -516,15 +511,10 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
   int phi_index;
   int k;
   int m;
-  int temp;
   int num_buses;
-  int T;
 
   // Num buses
   num_buses = NET_get_num_buses(CONSTR_get_network(c));
-
-  // Num periods
-  T = BRANCH_get_num_periods(br);
 
   // Constr data
   J = CONSTR_get_J(c);
@@ -534,7 +524,7 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
   bus_counted = CONSTR_get_bus_counted(c);
 
   // Check pointers
-  if (!J_nnz || !H_nnz || !bus_counted)
+  if (!J_nnz || !H_nnz || !H_array || !bus_counted)
     return;
 
   // Check outage
@@ -903,29 +893,6 @@ void CONSTR_ACPF_analyze_step(Constr* c, Branch* br, int t) {
 
     // Update counted flag
     bus_counted[bus_index_t[k]] = TRUE;
-  }
-
-  // Done
-  if ((t == T-1) && (BRANCH_get_index(br) == NET_get_num_branches(CONSTR_get_network(c))-1)) {
-
-    // Ensure lower triangular and save struct of H comb
-    H_nnz_comb = 0;
-    Hi_comb = MAT_get_row_array(CONSTR_get_H_combined(c));
-    Hj_comb = MAT_get_col_array(CONSTR_get_H_combined(c));
-    for (k = 0; k < CONSTR_get_H_array_size(c); k++) {
-      Hi = MAT_get_row_array(MAT_array_get(H_array,k));
-      Hj = MAT_get_col_array(MAT_array_get(H_array,k));
-      for (m = 0; m < MAT_get_nnz(MAT_array_get(H_array,k)); m++) {
-	if (Hi[m] < Hj[m]) {
-	  temp = Hi[m];
-	  Hi[m] = Hj[m];
-	  Hj[m] = temp;
-	}
-	Hi_comb[H_nnz_comb] = Hi[m];
-	Hj_comb[H_nnz_comb] = Hj[m];
-	H_nnz_comb++;
-      }
-    }
   }
 }