+++ src/problem/constr/constr_AC_FLOW_LIM.c
@@ -310,12 +310,7 @@ void CONSTR_AC_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
   int* H_nnz;
   int H_nnz_val;
   Mat* H_array;
-  Mat* H;
-  int* Hi;
-  int* Hj;
-  int* Hi_comb;
-  int* Hj_comb;
-  int H_nnz_comb;
+  Mat* H;    
   Mat* G; 
   Vec* l;
   Vec* u;
@@ -328,14 +323,11 @@ void CONSTR_AC_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
   int w_index[2];
   int a_index;
   int phi_index;
-  int temp;
   int k;
   int m;
-  int T;
   int num_vars;
   
-  // Num periods and net vars
-  T = BRANCH_get_num_periods(br);
+  // Num net vars
   num_vars = NET_get_num_vars(CONSTR_get_network(c));
 
   // Constr data
@@ -579,29 +571,6 @@ void CONSTR_AC_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
     
     // Constraint counter
     (*J_row)++;
-  }
-
-  // Done
-  if ((t == T-1) && (BRANCH_get_index(br) == NET_get_num_branches(CONSTR_get_network(c))-1)) {
-
-    // Ensure lower triangular and save struct of H comb
-    H_nnz_comb = 0;
-    Hi_comb = MAT_get_row_array(CONSTR_get_H_combined(c));
-    Hj_comb = MAT_get_col_array(CONSTR_get_H_combined(c));
-    for (k = 0; k < CONSTR_get_H_array_size(c); k++) {
-      Hi = MAT_get_row_array(MAT_array_get(H_array,k));
-      Hj = MAT_get_col_array(MAT_array_get(H_array,k));
-      for (m = 0; m < MAT_get_nnz(MAT_array_get(H_array,k)); m++) {
-	if (Hi[m] < Hj[m]) {
-	  temp = Hi[m];
-	  Hi[m] = Hj[m];
-	  Hj[m] = temp;
-	}
-	Hi_comb[H_nnz_comb] = Hi[m];
-	Hj_comb[H_nnz_comb] = Hj[m];
-	H_nnz_comb++;
-      }
-    }
   }
 }