+++ src/problem/constr/constr_REG_GEN.c
@@ -259,23 +259,17 @@ void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t) {
   Mat* H_array;
   Mat* Hy;
   Mat* Hz;
-  int* Hi;
-  int* Hj;
-  int* Hi_comb;
-  int* Hj_comb;
   int* A_nnz;
   int* J_nnz;
   int* A_row;
   int* J_row;
   int* H_nnz;
-  int H_nnz_comb;
   char* bus_counted;
   int bus_index_t[2];
   int index_y;
   int index_z;
   int k;
   int m;
-  int temp;
   int T;
   int num_vars;
 
@@ -296,7 +290,7 @@ void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t) {
   bus_counted = CONSTR_get_bus_counted(c);
 
   // Check pointers
-  if (!A_nnz || !J_nnz || !A_row || !J_row || !H_nnz || !bus_counted)
+  if (!A_nnz || !J_nnz || !A_row || !H_array || !J_row || !H_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -442,29 +436,6 @@ void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t) {
 
     // Update counted flag
     bus_counted[bus_index_t[k]] = TRUE;
-  }
-
-  // Done
-  if ((t == T-1) && (BRANCH_get_index(br) == NET_get_num_branches(CONSTR_get_network(c))-1)) {
-
-    // Ensure lower triangular and save struct of H comb
-    H_nnz_comb = 0;
-    Hi_comb = MAT_get_row_array(CONSTR_get_H_combined(c));
-    Hj_comb = MAT_get_col_array(CONSTR_get_H_combined(c));
-    for (k = 0; k < CONSTR_get_H_array_size(c); k++) {
-      Hi = MAT_get_row_array(MAT_array_get(H_array,k));
-      Hj = MAT_get_col_array(MAT_array_get(H_array,k));
-      for (m = 0; m < MAT_get_nnz(MAT_array_get(H_array,k)); m++) {
-	if (Hi[m] < Hj[m]) {
-	  temp = Hi[m];
-	  Hi[m] = Hj[m];
-	  Hj[m] = temp;
-	}
-	Hi_comb[H_nnz_comb] = Hi[m];
-	Hj_comb[H_nnz_comb] = Hj[m];
-	H_nnz_comb++;
-      }
-    }
   }
 }