+++ src/problem/constr/constr_REG_TRAN.c
@@ -253,19 +253,13 @@ void CONSTR_REG_TRAN_analyze_step(Constr* c, Branch* br, int tau) {
   Mat* Hvmax;
   Mat* Htmin;
   Mat* Htmax;
-  int* Hi;
-  int* Hj;
-  int* Hi_comb;
-  int* Hj_comb;
   int* A_nnz;
   int* J_nnz;
   int* A_row;
   int* J_row;
   int* H_nnz;
-  int H_nnz_comb;
   int k;
   int m;
-  int temp;
   int index_yz_vmin;
   int index_yz_vmax;
   int index_vvio_tmax;
@@ -295,7 +289,7 @@ void CONSTR_REG_TRAN_analyze_step(Constr* c, Branch* br, int tau) {
   H_nnz = CONSTR_get_H_nnz(c);
 
   // Check pointers
-  if (!A_nnz || !J_nnz || !A_row || !J_row || !H_nnz)
+  if (!A_nnz || !J_nnz || !A_row || !H_array || !J_row || !H_nnz)
     return;
 
   // Check outage
@@ -513,29 +507,6 @@ void CONSTR_REG_TRAN_analyze_step(Constr* c, Branch* br, int tau) {
     (*J_row)++; // CompVmax
     (*J_row)++; // CompTmax
     (*J_row)++; // CompTmin
-  }
-
-  // Done
-  if ((tau == T-1) && (BRANCH_get_index(br) == NET_get_num_branches(CONSTR_get_network(c))-1)) {
-    
-    // Ensure lower triangular and save struct of H comb
-    H_nnz_comb = 0;
-    Hi_comb = MAT_get_row_array(CONSTR_get_H_combined(c));
-    Hj_comb = MAT_get_col_array(CONSTR_get_H_combined(c));
-    for (k = 0; k < CONSTR_get_H_array_size(c); k++) {
-      Hi = MAT_get_row_array(MAT_array_get(H_array,k));
-      Hj = MAT_get_col_array(MAT_array_get(H_array,k));
-      for (m = 0; m < MAT_get_nnz(MAT_array_get(H_array,k)); m++) {
-	if (Hi[m] < Hj[m]) {
-	  temp = Hi[m];
-	  Hi[m] = Hj[m];
-	  Hj[m] = temp;
-	}
-	Hi_comb[H_nnz_comb] = Hi[m];
-	Hj_comb[H_nnz_comb] = Hj[m];
-	H_nnz_comb++;
-      }
-    }
   }
 }